requirements.txt
asgiref==3.8.1
Django>=5.2,<6.0
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.6,<2.10
pytz==2023.3
django-summernote>=0.8.20.0
gunicorn
python-decouple
whitenoise
pillow
python-dotenv>=1.1.1,<1.2
django-axes>=8.0.0,<8.1
unidecode>=1.4.0